// src/app/shared/models/customer-land-relation.model.ts
export class CustomerLandRelation {
  LandingProcedure: string;
  Cnic: string;
  LandInfoID: number;
  CCDetailId: number;
  BranchId: number;
  IsActive: string;
  IsOwner: string;
  CustomerName: string;
  CustomerNumber: string;
  PassbookNO: string;
  FatherName: string;
  EnteredBy: string;
  Status: string;
  Limit: string;
  Offset: string;
  TotalArea: string;
  LandCustID: number;
  ZoneId: number;
}